Revamp of the metrics folder
refactored naming of folder
deprecated wp-admin pages
relocated report functions to metrics.php
created draft critical path api, js, and functions
