+++ .gitignore
@@ -3,3 +3,4 @@
 /dt-people-groups/json/jp_pg_*.json
 /dt-locations/deprecated/
 /vendor/
+/dt-metrics/deprecated/
