+++ dt-metrics/class-counter-factory.php
