+++ dt-metrics/class-goals.php
@@ -1,0 +1,37 @@
+<?php
+
+/**
+ * Disciple_Tools_Goals
+ *
+ * @class Disciple_Tools_Goals
+ * @version 0.1
+ * @since 0.1
+ * @package Disciple_Tools
+ * @author Chasm.Solutions & Kingdom.Training
+ */
+
+if ( ! defined( 'ABSPATH' ) ) { exit; // Exit if accessed directly
+}
+
+/**
+ * Class Disciple_Tools_Goals
+ */
+class Disciple_Tools_Goals {
+
+    /**
+     * Constructor function.
+     * @access  public
+     * @since   0.1
+     */
+    public function __construct () {
+
+    } // End __construct()
+
+    /**
+     * @return bool
+     */
+    public function has_goals() {
+        return false;
+    }
+
+}
