+++ dt-notifications/notifications-template.php
@@ -12,3 +12,4 @@ if( !defined( 'ABSPATH' ) ) {
 } // Exit if accessed directly.
 
 /** Functions to output data for the theme. @see Buddypress bp-members-template.php or bp-groups-template.php for an example of the role of this page */
+
