--- dt-statistics/statistics-endpoints.php
@@ -1,17 +1,0 @@
-<?php
-/**
- * Custom endpoints file
- *
- * @package  Disciple_Tools
- * @category Plugin
- * @author   Chasm.Solutions & Kingdom.Training
- * @since    0.1
- */
-if ( ! defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly.
-
-/**
- * Class Disciple_Tools_Statistics_Endpoints
- */
-class Disciple_Tools_Statistics_Endpoints {
-
-}
