--- dt-statistics/statistics.php
@@ -1,1 +1,0 @@
-<?php
